Fixed issue when price with cents was rounded down.

+++ stripe-payments/public/includes/class-shortcode-asp.php
@@ -210,7 +210,7 @@ class AcceptStripePaymentsShortcode {
 	$button_id		 = 'stripe_button_' . $uniq_id;
 	$button_key		 = md5( uniqid() );
 	self::$payment_buttons[] = $button_id;
-	$paymentAmount		 = ($custom_quantity == "1" ? $price : (intval( $price ) * $quantity));
+	$paymentAmount		 = ($custom_quantity == "1" ? $price : (floatval( $price ) * $quantity));
 	if ( in_array( $currency, $this->AcceptStripePayments->zeroCents ) ) {
 	    //this is zero-cents currency, amount shouldn't be multiplied by 100
 	    $priceInCents = $paymentAmount;
